Improves Dimming Capability
If you give a simple dim/brighten [light], GA first needs to query the state.  Using the case method and created a function for this to populate JSON response correctly.  Then it will dim/brighten correctly based on the current state, and then set a new state using the existing adjustBrightness function.